[#53] refactor: 알림 페이지 리팩토링
